Major Organizational Overhaul
There's a lot to unpack here. I really cleaned up the css and the diretory structure of the site to get things more organized and orderly. Also I added a cool script to move around the window! The site is slowly but surely starting to really resemble an oldschool late 90s/early 2000s website!