Title Bar
Title Bar is looking good. Eventually I'm going to have to add some functionality and context menus to the buttons, but for now this is sufficient. Also I just realized the drag script works with the right mouse button too. Gotta fix that eventually. Javascript is not really my forte, so it'll probably be somewhere down the line. We'll see, but for now I still gotta implement the full GUI, so that's my current project.