ughhh
This is kind of a test. I've been tinkering around with stuff for a while. I think this will break the whole site but w/e. I basically reimplemented everything as iframes to make it easier to update and stuff, but I'm also trying to fix the javascript. We'll see what comes of this!